requirements.txt
Flask==3.1.1
flask_session==0.8.0
python-dotenv>=1.0.0
Flask-Mail>=0.10.0
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
plotly==5.24.1
Requests==2.32.3
seaborn==0.13.2
openpyxl==3.1.5
scipy==1.15.2
xgboost==2.1.4
joblib==1.4.2
scikit-learn>=1.3.0
# Аутентификация и безопасность
bcrypt>=4.1.0
# SSH-туннель для удалённого подключения к БД
paramiko>=3.4.0
sshtunnel>=0.4.0
# PostgreSQL (опционально, для продакшена)
# psycopg2-binary>=2.9.9
# Продакшен-сервер
gunicorn>=21.2.0
